Add new models: Test and related ones

+++ include/mind-net-api/mindnet/api/IPersistence.hpp
@@ -24,7 +24,7 @@ namespace mindnet::api
 
         virtual std::vector<std::string>& list_model_names() = 0;
         //
-        virtual std::pair<int, OperationResult> create(const ModelDefinition& def, AccessTokenContext& token,
+        virtual std::pair<i64, OperationResult> create(const ModelDefinition& def, AccessTokenContext& token,
                                                        entity_fields& fields) = 0;
         virtual std::pair<entity_fields, OperationResult> read(const ModelDefinition& def, AccessTokenContext& token,
                                                                i64 id) = 0;
